Logic CRUD done

+++ U5CDCG_HFT_2021221.Logic/CustomerLogic.cs
@@ -19,17 +19,46 @@ namespace U5CDCG_HFT_2021221.Logic
         }
         public void Create(Customer customer)
         {
-            CusRepo.Create(customer);
+            if (customer == null)
+            {
+                throw new NullReferenceException();
+            }
+            else if (customer.Age < 18)
+            {
+                throw new ArgumentOutOfRangeException("Must be older than 18");
+            }
+            else if (customer.Name == null)
+            {
+                throw new NullReferenceException();
+            }
+            else
+            {
+                CusRepo.Create(customer);
+            }
+           
         }
 
         public void Delete(int customerId)
         {
+            if (customerId <= 0)
+            {
+                throw new ArgumentOutOfRangeException();
+            }
             CusRepo.Delete(customerId);
         }
 
         public Customer Read(int customerId)
         {
-            return CusRepo.Read(customerId);
+            if (customerId <= 0)
+            {
+                throw new ArgumentOutOfRangeException();
+
+            }
+            else
+            {
+                return CusRepo.Read(customerId);
+            }
+            
         }
 
         public IQueryable<Customer> ReadAll()
@@ -39,7 +68,23 @@ namespace U5CDCG_HFT_2021221.Logic
 
         public void Update(Customer customer)
         {
-            CusRepo.Update(customer);
+
+            if (customer == null)
+            {
+                throw new NullReferenceException();
+            }
+            else if (customer.Age < 18)
+            {
+                throw new ArgumentOutOfRangeException("Must be older than 18");
+            }
+            else if (customer.Name == null)
+            {
+                throw new NullReferenceException();
+            }
+            else
+            {
+                CusRepo.Update(customer);
+            }
         }
     }
 }
